+++ apps/web/src/components/Explore/index.tsx
@@ -2,7 +2,7 @@ import MetaTags from '@components/Common/MetaTags';
 import RecommendedProfiles from '@components/Home/RecommendedProfiles';
 import Trending from '@components/Home/Trending';
 import Footer from '@components/Shared/Footer';
-import { GridItemEight, GridItemFour, GridLayout } from '@components/UI/GridLayout';
+import { GridItemEight, GridItemFour, GridLayout } from 'ui';
 import { Tab } from '@headlessui/react';
 import { Mixpanel } from '@lib/mixpanel';
 import { t } from '@lingui/macro';
