+++ apps/web/src/components/Messages/Message.tsx
@@ -1,7 +1,7 @@
 import MetaTags from '@components/Common/MetaTags';
 import MessageHeader from '@components/Messages/MessageHeader';
 import Loader from '@components/Shared/Loader';
-import { GridItemEight, GridLayout } from '@components/UI/GridLayout';
+import { GridItemEight, GridLayout } from 'ui';
 import useGetConversation from '@components/utils/hooks/useGetConversation';
 import useGetMessages from '@components/utils/hooks/useGetMessages';
 import useSendMessage from '@components/utils/hooks/useSendMessage';
