+++ apps/web/src/components/Profile/Shimmer.tsx
@@ -1,5 +1,5 @@
 import PublicationsShimmer from '@components/Shared/Shimmer/PublicationsShimmer';
-import { GridItemEight, GridItemFour, GridLayout } from '@components/UI/GridLayout';
+import { GridItemEight, GridItemFour, GridLayout } from 'ui';
 import type { FC } from 'react';
 
 const ProfilePageShimmer: FC = () => {
