+++ apps/web/src/components/Publication/Shimmer.tsx
@@ -2,7 +2,7 @@ import Footer from '@components/Shared/Footer';
 import PublicationShimmer from '@components/Shared/Shimmer/PublicationShimmer';
 import PublicationsShimmer from '@components/Shared/Shimmer/PublicationsShimmer';
 import UserProfileShimmer from '@components/Shared/Shimmer/UserProfileShimmer';
-import { GridItemEight, GridItemFour, GridLayout } from '@components/UI/GridLayout';
+import { GridItemEight, GridItemFour, GridLayout } from 'ui';
 import type { FC } from 'react';
 import { Card } from 'ui';
